Create webcrawler.rb
I have been working on getting information from LinkedIn. Step one is finding the names of some people working in a specified company, after that is done another task is writing a program that can do a google search of "Name Company LinkedIn" and subtract information from their LinkedIn page.

I have been able to get into a list of company employees at LinkedIn, and have done so by creating a new account as you can see in the code. The problem is that I can't see the names of people. This is because LinkedIn notices I am doing a lot of searches, and blocks my account for 24 hours! I need to rethink how to get around this..

I have an idea:
- Even if I am suspended I can see the top two results in a company page.
- By saving these names, and entering their profiles, you can continue clicking on "People Also Viewed" and hope you find people from the same company. The people in "People Also Viewed" ARE usually from the same company, and by doing some sort of if-statement you can only append their name if the company title matches.

I will continue working on this when I get back, feel free to have a look in the meantime.

OBS! You need some gems, as seen in the code to get it running