Total change of method
It is illegal to crawl with a LinkedIn account, and it didn't work well for me, so what I do now is:

1. Search on Google for "experience [company name] LinkedIn"
2. Find all links that look like "[something] | LinkedIn" scrolling through 10 google pages
3. Enter all these links and see if the person has [company name] under experience, and in that case, save all his/her skills with his/her name.

This method works, but you don't get that many people in the end. It's very sensitive to the country you input, as a search for a swedish company will be a lot better on swedish google than on the spanish google. Therefore I made country an input as well as company!

A big help would be to create a function that, once on a profile, clicks on profiles from "People Also Viewed" cause these are usually people from the same company. But this will work for now. 

A search with good hits is:
Company: Kidbrooke Advisory
Country: Sweden
(If you want to see how it works)

NOTE: Requires the file NameSkills too.