What’s new is that I now also return all the companies a person has worked for, because we might as well store this data on the spot.
On MongoLab there is a folder for each company AS THEY APPEAR ON
LINKEDIN, with the exception that I’ve transformed the names since
certain characters can’t be accepted.

This means a single person can appear in several folders now.

I’ve also done so that the crawler only runs if the company folder does
not exist. You can still decide to run if you feel you want a new
search to get more data.